the return value should be a 32-bit unsigned integer, not 64-bit

+++ src/math/summation.cxx
@@ -3,13 +3,13 @@
 namespace turbo {
 namespace math {
 
-std::uint64_t ceiling_bound(std::uint32_t base, std::uint32_t multiplier, std::uint32_t sum)
+std::uint32_t ceiling_bound(std::uint32_t base, std::uint32_t multiplier, std::uint32_t sum)
 {
     if (multiplier < 2U)
     {
 	throw invalid_multiplier_error("Multiplier cannot be less than 2");
     }
-    std::uint64_t log = 0U;
+    std::uint32_t log = 0U;
     std::uint64_t current = base;
     for (std::uint64_t tally = 0U; tally < sum;)
     {
